requirements.txt
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
nltk==3.4.5
python-dotenv==0.10.3
six==1.13.0
SQLAlchemy==1.3.11
textblob==0.15.3
Werkzeug==0.16.0
WTForms==2.2.1